+++ src/main/java/com/favoriteplace/app/controller/PostCommentController.java
@@ -12,7 +12,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.config.crypto.RsaKeyConversionServicePostProcessor;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
