+++ src/main/java/com/favoriteplace/app/service/community/CommentCommandService.java
@@ -2,8 +2,11 @@ package com.favoriteplace.app.service.community;
 
 import com.favoriteplace.app.domain.Member;
 import com.favoriteplace.app.domain.community.Comment;
+import com.favoriteplace.app.domain.community.GuestBook;
 import com.favoriteplace.app.domain.community.Post;
+import com.favoriteplace.app.dto.community.GuestBookRequestDto;
 import com.favoriteplace.app.repository.CommentRepository;
+import com.favoriteplace.app.repository.GuestBookRepository;
 import com.favoriteplace.app.repository.PostRepository;
 import com.favoriteplace.global.exception.ErrorCode;
 import com.favoriteplace.global.exception.RestApiException;
@@ -17,6 +20,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class CommentCommandService {
     private final PostRepository postRepository;
+    private final GuestBookRepository guestBookRepository;
     private final CommentRepository commentRepository;
 
     /**
@@ -64,6 +68,45 @@ public class CommentCommandService {
     }
 
     /**
+     * 성지순례 인증글에 댓글 추가
+     * @param member
+     * @param guestbookId
+     */
+    @Transactional
+    public void createGuestBookComment(Member member, Long guestbookId, GuestBookRequestDto.GuestBookCommentDto comment) {
+        GuestBook guestBook = guestBookRepository.findById(guestbookId).orElseThrow(() -> new RestApiException(ErrorCode.GUESTBOOK_NOT_FOUND));
+        Comment newComment = Comment.builder().member(member).guestBook(guestBook).content(comment.getContent()).build();
+        guestBook.addComment(newComment);
+        guestBookRepository.save(guestBook);
+    }
+
+    /**
+     * 성지순례 인증글에 댓글 수정
+     * @param member
+     * @param commentId
+     * @param content
+     */
+    @Transactional
+    public void modifyGuestBookComment(Member member, Long commentId, String content) {
+        Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new RestApiException(ErrorCode.COMMENT_NOT_FOUND));
+        checkAuthOfComment(member, comment);
+        Optional.ofNullable(content).ifPresent(comment::setContent);
+        commentRepository.save(comment);
+    }
+
+    /**
+     * 성지순례 인증글에 댓글 삭제
+     * @param member
+     * @param commentId
+     */
+    @Transactional
+    public void deleteGuestBookComment(Member member, Long commentId) {
+        Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new RestApiException(ErrorCode.COMMENT_NOT_FOUND));
+        checkAuthOfComment(member, comment);
+        commentRepository.delete(comment);
+    }
+
+    /**
      * 댓글의 작성자가 맞는지 확인하는 함수 (만약 작성자가 아니라면 에러 출력)
      * @param member
      * @param comment
@@ -73,4 +116,6 @@ public class CommentCommandService {
             throw new RestApiException(ErrorCode.USER_NOT_AUTHOR);
         }
     }
+
+
 }
